suspend: Detect and handle fan level reset on resume
On certain ThinkPad models like the T530, the fan level resets to "auto"
after system resume. In previous versions of zcfan this was generally
not a problem, as we'd always rewrite the fan level at the watchdog
threshold. However, in commit 0c3ac328d987 ("watchdog: Reset timer
instead of writing fan level to avoid spinup"), we started rewriting the
watchdog threshold instead of rewriting the fan level, which leaves the
fan level in "auto".

It's not ideal to simply revert this behaviour because 0c3ac328d987 too
fixes an issue where transitioning from level 0 to level 0 briefly spins
up the fans.

Introduce a mechanism to detect system resume using the difference
between CLOCK_MONOTONIC and CLOCK_BOOTTIME. CLOCK_MONOTONIC measures the
time since the system booted, excluding any time the system was
suspended, while CLOCK_BOOTTIME includes time spent in suspension. By
measuring and comparing the time deltas of both clocks over successive
checks, we can identify deviations caused by system suspension.

If the difference between the CLOCK_BOOTTIME delta and CLOCK_MONOTONIC
delta exceeds a threshold (0.2 seconds), a system resume is inferred.
The threshold is necessary to account for minor discrepancies due to
timing jitter or clock adjustments. In practice these are infrequent
enough that I'm not worried about 0 -> 0 transitions on rare events like
clock changes.

When a resume is detected, the current fan level is rewritten to ensure
it does not remain in "auto".

Fixes #35.